init: move fs related functions to its own file
